requirements.txt
streamlit
PyPDF2
requests
google-generativeai
pandas
numpy
